Address codex app PR feedback
